Update CLI interface
Adds a new parameter to run all puzzles, and provide a total duration
for running all puzzles.

Removes paramater to specify a file input. Instead the default is always
used.

With the default always being used, the library interface was updated a
little to combine the solutions with their input.