feat: optimize 2024-12-18 Part 2
Prior to this commit, day 18 would take ~250 ms.
Now, it takes only ~23 ms!
The secret? It's so simple I wish I had thought of it.
This speed up is thanks to binary searching for the first occurrence
where a path can't be made.
That's it.
I was able to delete a bit of code, and speed things up in the process.
Beautiful.