refactor: change git repo into a cargo workspace
The git repo was change into a cargo workspace to allow for multiple
crates.
This is to allow for possible rule changes for myself when writing
the puzzle solutions.
For example, 2024 was written to only use std, and with the goal of
running all solutions in under 1 second.
For 2025, or other years, I may change these rules.
To make the rules apparent, I decided that each year could be its own
crate.

In addition to this, I want to be able to write DRY'er code in two
areas:
- The puzzle runner
- Benchmarks

It irritated me to have to copy past a bunch in these situations.
Now that I can have multiple crates in the repository, that means
I can possibly make use of macros.